Move document sync into documents, fix some startup races
