+++ core/documents.py
@@ -1,4 +1,5 @@
 import sublime
+import sublime_plugin
 
 from collections import OrderedDict
 
@@ -12,8 +13,8 @@ from .logging import debug
 from .protocol import Notification, Point
 from .settings import settings
 from .url import filename_to_uri
-from .configurations import config_for_scope
-from .clients import client_for_view, window_clients
+from .configurations import config_for_scope, is_supported_view, is_supported_syntax, is_supportable_syntax
+from .clients import client_for_view, window_clients, check_window_unloaded
 from .events import Events
 
 SUBLIME_WORD_MASK = 515
@@ -172,6 +173,52 @@ def notify_did_change(view: sublime.View):
 
 
 document_sync_initialized = False
+
+
+class CloseListener(sublime_plugin.EventListener):
+    def on_close(self, view):
+        if is_supported_syntax(view.settings().get("syntax")):
+            Events.publish("view.on_close", view)
+        sublime.set_timeout_async(check_window_unloaded, 500)
+
+
+class SaveListener(sublime_plugin.EventListener):
+    def on_post_save_async(self, view):
+        if is_supported_view(view):
+            Events.publish("view.on_post_save_async", view)
+
+
+def is_transient_view(view):
+    window = view.window()
+    return view == window.transient_view_in_group(window.active_group())
+
+
+class DocumentSyncListener(sublime_plugin.ViewEventListener):
+    def __init__(self, view):
+        self.view = view
+
+    @classmethod
+    def is_applicable(cls, settings):
+        syntax = settings.get('syntax')
+        # This enables all of document sync for any supportable syntax
+        # Global performance cost, consider a detect_lsp_support setting
+        return syntax and (is_supported_syntax(syntax) or is_supportable_syntax(syntax))
+
+    @classmethod
+    def applies_to_primary_view_only(cls):
+        return False
+
+    def on_load_async(self):
+        # skip transient views: if not is_transient_view(self.view):
+        Events.publish("view.on_load_async", self.view)
+
+    def on_modified(self):
+        if self.view.file_name():
+            Events.publish("view.on_modified", self.view)
+
+    def on_activated_async(self):
+        if self.view.file_name():
+            Events.publish("view.on_activated_async", self.view)
 
 
 def initialize_document_sync(text_document_sync_kind):
